refactor(media-notification): Split routes out of controllers
media-service's mediaAsset.controller.ts and
notification-escalation-service's escalation.controller.ts and
notification.controller.ts move route/middleware declarations and
response Zod schemas into their own *.routes.ts files; controllers
keep only request-handling logic via createXController(service)
factories. module.ts files now create the shared DocumentedRouter and
register routes into it.

No behavior change — verified route-for-route identical OpenAPI doc
before/after for both services, lint clean, all 12 (media) and 16
(notification) existing tests pass, and live smoke checks against the
running services.
