+++ apps/notification-escalation-service/src/escalations/escalation.module.ts
@@ -1,14 +1,16 @@
-import type { DocumentedRouter } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { EscalationRepository } from './escalation.repository';
 import { EscalationService } from './escalation.service';
-import { createEscalationRouter } from './escalation.controller';
+import { registerEscalationRoutes } from './escalation.routes';
 
 /**
- * Composition root for the escalations feature: wires repository → service → router.
+ * Composition root for the escalations feature: wires repository → service → routes.
  */
 export function createEscalationModule(prisma: PrismaService): DocumentedRouter {
   const repository = new EscalationRepository(prisma);
   const service = new EscalationService(repository);
-  return createEscalationRouter(service);
+  const doc = createDocumentedRouter();
+  registerEscalationRoutes(doc, service);
+  return doc;
 }
