+++ apps/notification-escalation-service/src/notifications/notification.module.ts
@@ -1,16 +1,18 @@
-import type { DocumentedRouter } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { NotificationRepository } from './notification.repository';
 import { NotificationService } from './notification.service';
 import { SakhiClient } from './sakhi.client';
-import { createNotificationRouter } from './notification.controller';
+import { registerNotificationRoutes } from './notification.routes';
 
 /**
- * Composition root for the notifications feature: wires repository → service → router.
+ * Composition root for the notifications feature: wires repository → service → routes.
  * Replaces the former NestJS module + DI container.
  */
 export function createNotificationModule(prisma: PrismaService): DocumentedRouter {
   const repository = new NotificationRepository(prisma);
   const service = new NotificationService(repository, new SakhiClient());
-  return createNotificationRouter(service);
+  const doc = createDocumentedRouter();
+  registerNotificationRoutes(doc, service);
+  return doc;
 }
